Split migration-connector into modules
